[mlir][Python] move IRTypes and IRAttributes to public headers
