+++ mlir/examples/standalone/test/python/smoketest.py
@@ -23,6 +23,13 @@ with mlir_standalone.ir.Context():
     # CHECK: !standalone.custom<"foo">
     print(custom_type, file=sys.stderr)
 
+    # CHECK: this is a fp16 type
+    standalone_d.print_fp_type(mlir_standalone.ir.F16Type.get(), sys.stderr)
+    # CHECK: this is a fp32 type
+    standalone_d.print_fp_type(mlir_standalone.ir.F32Type.get(), sys.stderr)
+    # CHECK: this is a fp64 type
+    standalone_d.print_fp_type(mlir_standalone.ir.F64Type.get(), sys.stderr)
+
 
 # CHECK: Testing mlir package
 print("Testing mlir package", file=sys.stderr)
