+++ mlir/test/Examples/standalone/test.wheel.toy
@@ -39,6 +39,10 @@
 
 # CHECK: !standalone.custom<"foo">
 
+# CHECK: this is a fp16 type
+# CHECK: this is a fp32 type
+# CHECK: this is a fp64 type
+
 # CHECK: Testing mlir package
 
 # CHECK-NOT: RuntimeWarning: nanobind: type '{{.*}}' was already registered!
