+++ mlir/test/python/lib/PythonTestModuleNanobind.cpp
@@ -15,6 +15,7 @@
 #include "mlir-c/IR.h"
 #include "mlir/Bindings/Python/Diagnostics.h"
 #include "mlir/Bindings/Python/IRCore.h"
+#include "mlir/Bindings/Python/IRTypes.h"
 #include "mlir/Bindings/Python/Nanobind.h"
 #include "mlir/Bindings/Python/NanobindAdaptors.h"
 #include "nanobind/nanobind.h"
@@ -47,6 +48,49 @@ struct PyTestType
   }
 };
 
+struct PyTestIntegerRankedTensorType
+    : mlir::python::MLIR_BINDINGS_PYTHON_DOMAIN::PyConcreteType<
+          PyTestIntegerRankedTensorType,
+          mlir::python::MLIR_BINDINGS_PYTHON_DOMAIN::PyRankedTensorType> {
+  static constexpr IsAFunctionTy isaFunction = mlirTypeIsARankedIntegerTensor;
+  static constexpr GetTypeIDFunctionTy getTypeIdFunction =
+      mlirRankedTensorTypeGetTypeID;
+  static constexpr const char *pyClassName = "TestIntegerRankedTensorType";
+  using PyConcreteType::PyConcreteType;
+
+  static void bindDerived(ClassTy &c) {
+    c.def_static(
+        "get",
+        [](std::vector<int64_t> shape, unsigned width,
+           mlir::python::MLIR_BINDINGS_PYTHON_DOMAIN::DefaultingPyMlirContext
+               ctx) {
+          MlirAttribute encoding = mlirAttributeGetNull();
+          return PyTestIntegerRankedTensorType(
+              ctx->getRef(),
+              mlirRankedTensorTypeGet(
+                  shape.size(), shape.data(),
+                  mlirIntegerTypeGet(ctx.get()->get(), width), encoding));
+        },
+        nb::arg("shape"), nb::arg("width"),
+        nb::arg("context").none() = nb::none());
+  }
+};
+
+struct PyTestTensorValue
+    : mlir::python::MLIR_BINDINGS_PYTHON_DOMAIN::PyConcreteValue<
+          PyTestTensorValue> {
+  static constexpr IsAFunctionTy isaFunction =
+      mlirTypeIsAPythonTestTestTensorValue;
+  static constexpr GetTypeIDFunctionTy getTypeIdFunction =
+      mlirRankedTensorTypeGetTypeID;
+  static constexpr const char *pyClassName = "TestTensorValue";
+  using PyConcreteValue::PyConcreteValue;
+
+  static void bindDerived(ClassTy &c) {
+    c.def("is_null", [](MlirValue &self) { return mlirValueIsNull(self); });
+  }
+};
+
 class PyTestAttr
     : public mlir::python::MLIR_BINDINGS_PYTHON_DOMAIN::PyConcreteAttribute<
           PyTestAttr> {
@@ -73,18 +117,18 @@ public:
 NB_MODULE(_mlirPythonTestNanobind, m) {
   m.def(
       "register_python_test_dialect",
-      [](MlirContext context, bool load) {
+      [](mlir::python::MLIR_BINDINGS_PYTHON_DOMAIN::DefaultingPyMlirContext
+             context,
+         bool load) {
         MlirDialectHandle pythonTestDialect =
             mlirGetDialectHandle__python_test__();
-        mlirDialectHandleRegisterDialect(pythonTestDialect, context);
+        mlirDialectHandleRegisterDialect(pythonTestDialect,
+                                         context.get()->get());
         if (load) {
-          mlirDialectHandleLoadDialect(pythonTestDialect, context);
+          mlirDialectHandleLoadDialect(pythonTestDialect, context.get()->get());
         }
       },
-      nb::arg("context"), nb::arg("load") = true,
-      // clang-format off
-      nb::sig("def register_python_test_dialect(context: " MAKE_MLIR_PYTHON_QUALNAME("ir.Context") ", load: bool = True) -> None"));
-  // clang-format on
+      nb::arg("context").none() = nb::none(), nb::arg("load") = true);
 
   m.def(
       "register_dialect",
@@ -100,73 +144,16 @@ NB_MODULE(_mlirPythonTestNanobind, m) {
 
   m.def(
       "test_diagnostics_with_errors_and_notes",
-      [](MlirContext ctx) {
-        mlir::python::CollectDiagnosticsToStringScope handler(ctx);
-        mlirPythonTestEmitDiagnosticWithNote(ctx);
+      [](mlir::python::MLIR_BINDINGS_PYTHON_DOMAIN::DefaultingPyMlirContext
+             ctx) {
+        mlir::python::CollectDiagnosticsToStringScope handler(ctx.get()->get());
+        mlirPythonTestEmitDiagnosticWithNote(ctx.get()->get());
         throw nb::value_error(handler.takeMessage().c_str());
       },
-      // clang-format off
-      nb::sig("def test_diagnostics_with_errors_and_notes(arg: " MAKE_MLIR_PYTHON_QUALNAME("ir.Context") ", /) -> None"));
-  // clang-format on
+      nb::arg("context").none() = nb::none());
 
   PyTestAttr::bind(m);
   PyTestType::bind(m);
-
-  auto typeCls =
-      mlir_type_subclass(m, "TestIntegerRankedTensorType",
-                         mlirTypeIsARankedIntegerTensor,
-                         nb::module_::import_(MAKE_MLIR_PYTHON_QUALNAME("ir"))
-                             .attr("RankedTensorType"))
-          .def_classmethod(
-              "get",
-              [](const nb::object &cls, std::vector<int64_t> shape,
-                 unsigned width, MlirContext ctx) {
-                MlirAttribute encoding = mlirAttributeGetNull();
-                return cls(mlirRankedTensorTypeGet(
-                    shape.size(), shape.data(), mlirIntegerTypeGet(ctx, width),
-                    encoding));
-              },
-              // clang-format off
-              nb::sig("def get(cls: object, shape: collections.abc.Sequence[int], width: int, context: " MAKE_MLIR_PYTHON_QUALNAME("ir.Context") " | None = None) -> object"),
-              // clang-format on
-              nb::arg("cls"), nb::arg("shape"), nb::arg("width"),
-              nb::arg("context").none() = nb::none());
-
-  assert(nb::hasattr(typeCls.get_class(), "static_typeid") &&
-         "TestIntegerRankedTensorType has no static_typeid");
-
-  MlirTypeID mlirRankedTensorTypeID = mlirRankedTensorTypeGetTypeID();
-
-  nb::module_::import_(MAKE_MLIR_PYTHON_QUALNAME("ir"))
-      .attr(MLIR_PYTHON_CAPI_TYPE_CASTER_REGISTER_ATTR)(
-          mlirRankedTensorTypeID, nb::arg("replace") = true)(
-          nanobind::cpp_function([typeCls](const nb::object &mlirType) {
-            return typeCls.get_class()(mlirType);
-          }));
-
-  auto valueCls = mlir_value_subclass(m, "TestTensorValue",
-                                      mlirTypeIsAPythonTestTestTensorValue)
-                      .def("is_null", [](MlirValue &self) {
-                        return mlirValueIsNull(self);
-                      });
-
-  nb::module_::import_(MAKE_MLIR_PYTHON_QUALNAME("ir"))
-      .attr(MLIR_PYTHON_CAPI_VALUE_CASTER_REGISTER_ATTR)(
-          mlirRankedTensorTypeID)(
-          nanobind::cpp_function([valueCls](const nb::object &valueObj) {
-            std::optional<nb::object> capsule =
-                mlirApiObjectToCapsule(valueObj);
-            assert(capsule.has_value() && "capsule is not null");
-            MlirValue v = mlirPythonCapsuleToValue(capsule.value().ptr());
-            MlirType t = mlirValueGetType(v);
-            // This is hyper-specific in order to exercise/test registering a
-            // value caster from cpp (but only for a single test case; see
-            // testTensorValue python_test.py).
-            if (mlirShapedTypeHasStaticShape(t) &&
-                mlirShapedTypeGetDimSize(t, 0) == 1 &&
-                mlirShapedTypeGetDimSize(t, 1) == 2 &&
-                mlirShapedTypeGetDimSize(t, 2) == 3)
-              return valueCls.get_class()(valueObj);
-            return valueObj;
-          }));
+  PyTestIntegerRankedTensorType::bind(m);
+  PyTestTensorValue::bind(m);
 }
